Improve publishing steps. Add formal LICENSE. Improve documentation. Add TypeScript definitions.
