+++ index.d.ts
@@ -1,10 +1,40 @@
-declare module 'react-ssr-prepass' {
-  type Visitor = (
-    element: React.ElementType<any>,
-    instance?: React.Component<any, any>,
-  ) => void | Promise<any>;
+import React from 'react';
+import { NextComponentClass, NextContext } from 'next';
+import { Client, ClientOptions, Exchange } from 'urql';
+import { SSRExchange, SSRData } from 'urql/dist/types/exchanges/ssr';
 
-  function ssrPrepass(node: React.ReactNode, visitor?: Visitor): Promise<void>;
-
-  export = ssrPrepass;
+interface WithUrqlClient {
+  urqlClient: Client;
 }
+
+interface WithUrqlState {
+  urqlState: SSRData;
+}
+
+interface NextContextWithAppTree extends NextContext {
+  AppTree: React.ComponentType<any>;
+}
+
+declare const withUrqlClient: <T extends {}>(
+  clientOptions: Pick<
+    ClientOptions,
+    'fetch' | 'url' | 'fetchOptions' | 'requestPolicy'
+  >,
+  mergeExchanges?: (ssrEx: SSRExchange) => Exchange[],
+) => (
+  App:
+    | NextComponentClass<
+        T & WithUrqlClient,
+        T & WithUrqlClient,
+        NextContext<Record<string, string | string[] | undefined>, {}>
+      >
+    | import('next').NextFunctionComponent<
+        T & WithUrqlClient,
+        T & WithUrqlClient,
+        NextContext<Record<string, string | string[] | undefined>, {}>
+      >,
+) => import('next').NextFunctionComponent<
+  T & WithUrqlClient & WithUrqlState,
+  T & WithUrqlState,
+  NextContextWithAppTree
+>;
