+++ package.json
@@ -9,14 +9,17 @@
   "main": "dist/next-urql.min.js",
   "module": "dist/next-urql.es.min.js",
   "source": "src/index.ts",
+  "types": "index.d.ts",
   "author": "Formidable",
   "license": "MIT",
   "scripts": {
+    "build": "rollup -c rollup.config.js",
     "check:types": "tsc",
+    "clean": "rimraf ./dist ./node_modules/.cache",
     "format": "prettier --write **/*.{ts,tsx,d.ts,js}",
     "lint": "eslint **/*.{ts,tsx}",
-    "test": "jest",
-    "build": "rollup -c rollup.config.js"
+    "prepublishOnly": "run-s clean test build",
+    "test": "jest"
   },
   "devDependencies": {
     "@babel/core": "^7.6.4",
@@ -38,10 +41,12 @@
     "graphql-tag": "^2.10.1",
     "isomorphic-unfetch": "^3.0.0",
     "jest": "^24.9.0",
+    "npm-run-all": "^4.1.5",
     "prettier": "^1.18.2",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-is": "^16.10.2",
+    "rimraf": "^3.0.0",
     "rollup": "^1.25.0",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-commonjs": "^10.1.0",
