+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
   "compilerOptions": {
     "baseUrl": ".",
+    "declaration": true,
     "esModuleInterop": true,
     "jsx": "react",
     "lib": ["dom", "esnext"],
     "noUnusedLocals": true,
-    "outDir": "dist/cjs",
     "rootDir": "./src",
     "module": "es2015",
     "moduleResolution": "node",
@@ -16,5 +16,5 @@
     "strict": true,
     "target": "esnext"
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx", "index.d.ts"]
+  "include": ["src/**/*.ts", "src/**/*.tsx", "**/*.d.ts"]
 }
